pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "home_financial_tools"
description = "Home Financial Tools"
requires-python = ">=3.12"
authors = [{ name = "Zhuo Yin", email = "zhuoyin@gmail.com" }]
dependencies = ["python-dateutil", "fpdf2", "fastapi", "uvicorn", "pyyaml", "pydantic", "bcrypt", "slowapi"]
dynamic = ["version"]

[project.scripts]
gen_invoice = "home_financial_tools.entry_points.invoice:main"

[tool.setuptools]
packages = ["home_financial_tools"]

[tool.setuptools.dynamic]
version = {attr = "home_financial_tools.__version__"}
